Refactor: schedule page for streaming #108
